+++ app/src/main/java/com/concordium/wallet/ui/cis2/TokenItemSeparatorDecorator.kt
@@ -1,0 +1,24 @@
+package com.concordium.wallet.ui.cis2
+
+import android.graphics.Canvas
+import android.graphics.drawable.Drawable
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class TokenItemSeparatorDecorator(private val divider: Drawable?) : RecyclerView.ItemDecoration() {
+
+    override fun onDrawOver(canvas: Canvas, parent: RecyclerView, state: RecyclerView.State) {
+        val dividerLeft = parent.paddingLeft
+        val dividerRight = parent.width - parent.paddingRight
+        val childCount = parent.childCount
+        for (i in 0..childCount - 2) {
+            val child: View = parent.getChildAt(i)
+            val params =
+                child.layoutParams as RecyclerView.LayoutParams
+            val dividerTop: Int = child.bottom + params.bottomMargin
+            val dividerBottom = dividerTop + (divider?.intrinsicHeight?:0)
+            divider?.setBounds(dividerLeft, dividerTop, dividerRight, dividerBottom)
+            divider?.draw(canvas)
+        }
+    }
+}